Basic build & publish pipeline
