+++ src/index.ts
@@ -1,0 +1,3 @@
+import initTerminal from './api/init'
+
+export { initTerminal }
